# Improved performance of comparison with SIMD feature flag (2x-3.5x)