--- a/src/compute/comparison/simd/packed.rs
+++ b/src/compute/comparison/simd/packed.rs
@@ -1,3 +1,5 @@
+use std::convert::TryInto;
+
 use super::{set, Simd8, Simd8Lanes};
 
 use packed_simd::*;