@jorgecarleitao 2021-08-20
There is an extra optimization in using the more natural md representation of the native types, but it requires a bit more work on the `MutableBitmap` to be safely constructed from `u8-u64` chunks